+++ commands/punnish.js
@@ -32,7 +32,9 @@ module.exports.run = async (bot, message, args) => {
               //message.channel.send({content: "Witch punnishment action do you wana do?", components: [row] });
               message.channel.send({content: "Command Disabled", components: [row] });
               
-			  message.delete();
+			  if (message.guild.me.permissions.has("MANAGE_MESSAGES")){
+				message.delete();
+				}
 }
 
 module.exports.help = {
